+++ src/libraries/tools/SocketManager.h
@@ -50,7 +50,7 @@ public:
     //@{
     virtual unsigned long GetNbMessagesSent() const { return nbMessagesSent_; }
     virtual bool HasAnsweredSinceLastTick( const std::string& endpoint );
-    size_t GetSentAmount() const;
+    uint64_t GetSentAmount() const;
     //@}
 
 private:
@@ -75,7 +75,7 @@ private:
     const DWORD timeOut_;
     T_Sockets sockets_;
     unsigned long nbMessagesSent_;
-    size_t outBytes_;
+    uint64_t outBytes_;
     //@}
 };
